Added Words particle system for SplashScene.
I designed and built this entire particle system on a test bed and
transferred it over today. It's in letters.h and letters.cpp.
This is a big change. The particle system is an "instanced"
particle system, thus giving it faster rendering speed.  This also
uses Bezier curves for letter paths.  The letters spell out
"PHYSICS SIMULATIONS".

It replaces the letters in circles orbiting
points from the prior SplashScene, which was in objectunit.cpp/.h
and particlesystem.cpp/.h.  This scene was too slow with the
particle system. It also spelled out "PHYSICS SIMULATIONS", but
with lettered textures, not particles.